Adding the baseline model files
The set contains all the files needed to run baseline experiments. Currently set with the same hyper parameters as the original. Further explanation in the readme file.
Note: First step is to run mat2py file on the data. Following that, run the 'ipynb' on the same env for training.
Encountering a bug in main.py, looking into it.